Add files via upload
Welcome to the Zomato Clone repository! 🚀
I Shubham Singh Parihar the create of this project.

This project is a front-end clone of Zomato's website, created using HTML and CSS. The primary goal of this project was to practice web development skills and demonstrate proficiency in crafting responsive, visually appealing layouts using basic web technologies.